+++ server/src/main/java/edu/colostate/cs/cs414/chesshireCoders/jungleServer/daos/GamePieceDAO.java
@@ -1,8 +1,7 @@
-package edu.colostate.cs.cs414.chesshireCoders.jungleServer.dataAccessObjects;
+package edu.colostate.cs.cs414.chesshireCoders.jungleServer.daos;
 
 import edu.colostate.cs.cs414.chesshireCoders.jungleUtil.types.PieceType;
 import edu.colostate.cs.cs414.chesshireCoders.jungleUtil.game.GamePiece;
-import edu.colostate.cs.cs414.chesshireCoders.jungleServer.server.JungleDB;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
