+++ server/src/main/java/edu/colostate/cs/cs414/chesshireCoders/jungleServer/session/LoginVerifier.java
@@ -1,7 +1,7 @@
 package edu.colostate.cs.cs414.chesshireCoders.jungleServer.session;
 
-import edu.colostate.cs.cs414.chesshireCoders.jungleServer.dataAccessObjects.LoginDAO;
-import edu.colostate.cs.cs414.chesshireCoders.jungleServer.dataAccessObjects.UserDAO;
+import edu.colostate.cs.cs414.chesshireCoders.jungleServer.daos.LoginDAO;
+import edu.colostate.cs.cs414.chesshireCoders.jungleServer.daos.UserDAO;
 import edu.colostate.cs.cs414.chesshireCoders.jungleServer.server.JungleDB;
 import edu.colostate.cs.cs414.chesshireCoders.jungleUtil.game.Login;
 import edu.colostate.cs.cs414.chesshireCoders.jungleUtil.game.User;
